+++ src/main/scala/io/opentargets/etl/backend/spark/Helpers.scala
@@ -1,26 +1,42 @@
 package io.opentargets.etl.backend.spark
 
-import io.opentargets.etl.backend.ETLSessionContext
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.spark.SparkConf
 import org.apache.spark.sql._
-import org.apache.spark.sql.functions._
+import org.apache.spark.sql.functions.{
+  array,
+  array_distinct,
+  array_union,
+  coalesce,
+  col,
+  expr,
+  filter,
+  flatten,
+  lit,
+  struct,
+  substring_index,
+  typedLit
+}
 import org.apache.spark.sql.types.{ArrayType, DataType, StructField, StructType}
 
 import scala.language.postfixOps
 
-import monocle.macros.syntax.lens._
-
 object Helpers extends LazyLogging {
 
-  case class Metadata(id: String,
-                      resource: IOResourceConfig,
-                      serialisedSchema: String,
-                      columns: List[String])
+  /** Returns input string wrapped in backticks if it contains period character.
+    *
+    * Spark interprets the . symbol to be a select. Input files may include this in their column names causing
+    * unanticipated behaviour.
+    */
+  val wrapColumnNamesWithPeriodCharacters: String => String = {
+    case a if a.contains(".") => s"`$a`"
+    case s                    => s
+  }
 
   /** generate a spark session given the arguments if sparkUri is None then try to get from env
     * otherwise it will set the master explicitely
-    * @param appName the app name
+    *
+    * @param appName  the app name
     * @param sparkUri uri for the spark env master if None then it will try to get from yarn
     * @return a sparksession object
     */
@@ -97,148 +113,6 @@ object Helpers extends LazyLogging {
             |t -> isnotnull(t))""".stripMargin)
   }
 
-  type WriterConfigurator = DataFrameWriter[Row] => DataFrameWriter[Row]
-
-  /** It creates an hashmap of dataframes.
-    *   Es. inputsDataFrame {"disease", Dataframe} , {"target", Dataframe}
-    *   Reading is the first step in the pipeline
-    */
-  def readFrom(
-      inputFileConf: IOResourceConfigurations
-  )(implicit session: SparkSession): IOResources = {
-    logger.info("Load files into a Map of names and IOResource")
-    for {
-      (key, formatAndPath) <- inputFileConf
-    } yield key -> IOResource(loadFileToDF(formatAndPath), formatAndPath)
-  }
-
-  def loadFileToDF(pathInfo: IOResourceConfig)(implicit session: SparkSession): DataFrame = {
-    logger.info(s"load dataset ${pathInfo.path} with ${pathInfo.toString}")
-
-    pathInfo.options
-      .foldLeft(session.read.format(pathInfo.format)) { (dfr, opts) =>
-        val options = opts.map(c => c.k -> c.v).toMap
-        dfr.options(options)
-      }
-      .load(pathInfo.path)
-  }
-
-  /**
-    * Helper function to prepare multiple files of the same category to be read by `readFrom`
-    * @param resourceConfigs collection of IOResourceConfig of unknown composition
-    * @return Map with random keys to input resource.
-    */
-  def seqToIOResourceConfigMap(resourceConfigs: Seq[IOResourceConfig]): IOResourceConfigurations = {
-    (for (rc <- resourceConfigs) yield scala.util.Random.alphanumeric.take(6).toString -> rc).toMap
-  }
-
-  /**
-    * Given an IOResource ior and the metadata config section it generates a one-line DF that
-    * will be saved coalesced to 1 into a folder inside the metadata output folder. This will
-    * make easier to collect the matadata of the created resources
-    * @param ior the IOResource from which generates the metadata
-    * @param withConfig the metadata Config section
-    * @param context ETL context object
-    * @return a new IOResource with all needed information and data ready to be saved
-    */
-  private def generateMetadata(ior: IOResource, withConfig: IOResourceConfig)(
-      implicit context: ETLSessionContext): IOResource = {
-    require(!withConfig.path.isBlank, "metadata resource path cannot be empty")
-    implicit val session: SparkSession = context.sparkSession
-    import session.implicits._
-
-    val serialisedSchema = ior.data.schema.json
-    val iores =
-      ior.configuration
-        .lens(_.path)
-        .modify(
-          _.stripPrefix(context.configuration.common.output)
-            .split("/")
-            .filterNot(_.isBlank)
-            .mkString("/", "/", ""))
-    val cols = ior.data.columns.toList
-    val id = ior.configuration.path.split("/").filterNot(_.isBlank).last
-    val newPath = withConfig.path + s"/$id"
-    val metadataConfig = withConfig.lens(_.path).set(newPath)
-
-    val metadata =
-      List(Metadata(id, iores, serialisedSchema, cols)).toDF
-        .withColumn("timeStamp", current_timestamp())
-        .coalesce(numPartitions = 1)
-
-    val metadataIOResource = IOResource(metadata, metadataConfig)
-
-    logger.info(s"generate metadata info for $id in path $newPath")
-
-    metadataIOResource
-  }
-
-  private def writeTo(output: IOResource)(implicit context: ETLSessionContext): IOResource = {
-    implicit val spark: SparkSession = context.sparkSession
-
-    logger.info(s"save IOResource ${output.toString}")
-    val data = output.data
-    val conf = output.configuration
-
-    val pb = conf.partitionBy.foldLeft(data.write) {
-      case (df, ops) =>
-        logger.debug(s"enabled partition by ${ops.toString}")
-        df.partitionBy(ops: _*)
-    }
-
-    conf.options
-      .foldLeft(pb) {
-        case (df, ops) =>
-          logger.debug(s"write to ${conf.path} with options ${ops.toString}")
-          val options = ops.map(c => c.k -> c.v).toMap
-          df.options(options)
-
-      }
-      .format(conf.format)
-      .save(conf.path)
-
-    output
-  }
-
-  /**
-    * writeTo save all datasets in the Map outputs. It does write per IOResource
-    * its companion metadata dataset
-    * @param outputs the Map with all IOResource
-    * @param context the context to have the configuration and the spark session
-    * @return the same outputs as a continuator
-    */
-  def writeTo(outputs: IOResources)(implicit context: ETLSessionContext): IOResources = {
-    implicit val spark: SparkSession = context.sparkSession
-
-    val datasetNamesStr = outputs.keys.mkString("(", ", ", ")")
-    logger.info(s"write datasets $datasetNamesStr")
-
-    outputs foreach { out =>
-      logger.info(s"save dataset ${out._1}")
-      writeTo(out._2)
-
-      logger.info(s"save metadata for dataset ${out._1}")
-      val md = generateMetadata(out._2, context.configuration.common.metadata)
-      writeTo(md)
-    }
-
-    outputs
-  }
-
-  /** generate a set of String with the union of Columns.
-    * Eg, myCols =( a,c,d) and allCols(a,c,d,e,f,h)
-    * return (a,c,d,e,f,h)
-    * @param myCols the list of the Columns in a specific Dataframe
-    * @param allCols the list of Columns to match
-    * @return a sparksession object
-    */
-  def columnExpr(myCols: Set[String], allCols: Set[String]): Set[Column] = {
-    val inter = (allCols intersect myCols).map(col)
-    val differ = (allCols diff myCols).map(lit(null).as(_))
-
-    inter union differ
-  }
-
   /** generate the union between two dataframe with different Schema.
     * df is the implicit dataframe
     *
@@ -290,18 +164,6 @@ object Helpers extends LazyLogging {
     newDF
   }
 
-  // Replace the spaces from the schema fields with _
-  def replaceSpacesSchema(df: DataFrame)(implicit session: SparkSession): DataFrame = {
-
-    //replace all spaces with _
-    val renameFcn = (s: String) => s.replaceAll(" ", "_")
-
-    val newDF =
-      session.createDataFrame(df.rdd, renameAllCols(df.schema, renameFcn))
-
-    newDF
-  }
-
   def renameAllCols(schema: StructType, fn: String => String): StructType = {
 
     def renameDataType(dt: StructType): StructType =
@@ -319,12 +181,25 @@ object Helpers extends LazyLogging {
     renameDataType(schema)
   }
 
+  // Replace the spaces from the schema fields with _
+  def replaceSpacesSchema(df: DataFrame)(implicit session: SparkSession): DataFrame = {
+
+    //replace all spaces with _
+    val renameFcn = (s: String) => s.replaceAll(" ", "_")
+
+    val newDF =
+      session.createDataFrame(df.rdd, renameAllCols(df.schema, renameFcn))
+
+    newDF
+  }
+
   /** Given a dataframe with a n columns, this method create a new column called `collectUnder` which will include all
     * columns listed in `includedColumns` in a struct column. Those columns will be removed from the original dataframe.
     * This can be used to nest fields.
-    * @param dataFrame on which to perform nesting
+    *
+    * @param dataFrame       on which to perform nesting
     * @param includedColumns columns to include in new nested column
-    * @param collectUnder name of new struct column
+    * @param collectUnder    name of new struct column
     * @return dataframe with new column `collectUnder` with `includedColumns` nested within it.
     */
   def nest(dataFrame: DataFrame, includedColumns: List[String], collectUnder: String): DataFrame = {
@@ -362,13 +237,4 @@ object Helpers extends LazyLogging {
     columns.map(coalesce(_, typedLit(Array.empty))).reduce((c1, c2) => array_union(c1, c2))
   }
 
-  /** Returns input string wrapped in backticks if it contains period character.
-    *
-    * Spark interprets the . symbol to be a select. Input files may include this in their column names causing
-    * unanticipated behaviour.
-    */
-  val wrapColumnNamesWithPeriodCharacters: String => String = {
-    case a if a.contains(".") => s"`$a`"
-    case s                    => s
-  }
 }
